Predict function for srlTS object

Predict function for srlTS object

Usage

# S3 method for srlTS
predict(object, n_ahead = 1, X_test, y_test, cumulative = 0, ...)

Arguments

object

an srlTS object

n_ahead

number of times ahead to predict by iteration

X_test

a matrix exogenous features

y_test

the test series; needed for future predictions (optional; see details)

cumulative

should cumulative (rolling) sums be returned (integer indicating number of times to sum)

...

currently unused

Value

a vector of predictions

Details

The `y_test` argument must be supplied if forecasts are desired or if `n_ahead` < `nrow(X_test)`. This is because in order to obtain 1-step forecast for, say, the 10th observation in the test data set, the 9th observation of `y_test` is required. The length of `y_test` will determine how many forecasts to produce. In order to get true forecasts for the first 30 observations after the training set, one must (currently) produce the set of 1-step, 2-step, 3-step, ..., 30-step ahead predictions.

Examples

data("LakeHuron")
fit_LH <- srlTS(LakeHuron)
predict(fit_LH)
#>  [1]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [9]       NA 581.0125 580.7219 580.9960 581.1889 580.5357 580.1159 579.7550
#> [17] 579.7761 578.8827 579.2512 579.5325 579.4836 578.1434 578.5270 579.3470
#> [25] 578.8931 579.3384 578.6304 579.4781 578.8051 579.0307 579.8600 579.5126
#> [33] 579.5211 579.7656 579.7974 579.0669 578.6431 578.2998 578.9424 579.5938
#> [41] 578.5703 578.1132 579.8013 579.8827 579.7640 579.2026 579.2054 578.5818
#> [49] 579.0279 577.9377 578.0709 576.9041 577.3496 578.3401 578.7360 580.6749
#> [57] 578.6937 577.2520 577.5200 577.4140 576.5386 577.5373 577.1386 577.2966
#> [65] 578.2541 578.2710 577.5091 577.5881 578.9097 579.6057 578.6401 579.3499
#> [73] 579.1059 579.3386 578.9476 577.8435 578.5245 580.0310 580.5006 579.7886
#> [81] 579.6993 579.4481 578.5976 578.2988 577.3532 577.6201 579.6514 577.7694
#> [89] 578.1514 577.0373 576.4356 577.6168 578.0205 578.5252 578.5060 579.9080
#> [97] 578.8826 579.9434